Add files via upload
The CRUD (Create,read,Update,Delete)project involved creating and managing a database of employee record table. Settings file contains all the configuration of  Django installation. 
urls.py use - a request in Django first comes to urls.py and then goes to the matching function in views.py. 
views.py - Python functions in views.py take the web request from urls.py and give the web response to templates. -Templat is the presentation layer.
models.py --It contains the essential fields and behaviors of the data  storing. Generally, each model maps to a single database table.
add.html - it contains data of the employee and  add the data of the tableusing HTML5,CSS3,Django
indexc.html-It show the whole data of the table and search tha particular data of employee.
update.html- it update the data of the  employee  in  the table